Support for Split operation
We support splitting of a tensor along an axis into n pieces, where n
has to be a constant.
Eg:
  Split(Tensor of shape(5,30), splits=3, axis=1)
  returns 3 tensors of shape(5,10) each.

Currently we do not suport splitting into tensors of specified shape
(num_or_size_splits) though that functionality will be added later.

We also do not support splitting into n pieces where n is a runtime
value because we do not support run-time code generation yet.

This also adds support in the frontend for an op to return multiple
values.