Pass key directory as argument for Porthos.
Previously programs linked with porthos, expected the keys to be in the
current_directory/files folder. Now users can explicitly pass the directory
while invoking the program. The keys still have to be named keyA, keyAB, keyB
and keyD but they can be located in arbitary folders.

So party 0 will invoke the program as:
  ./program 0 files/addresses_file path/to/keys/dir