[ONNX] Add feature to strip model weights
Now we can compile models similar to how we compile tensorflow models. The
server can compile the model with:
  ~/EzPC/Athos/CompileONNXGraph.py --config config.json --role server

This generates a zip file which contains:
  optimised_model.onnx
  config.json

The optimised_model.onnx is the same model but is a pruned version and doesnt
contain model weights. This zip file can then be sent to the client and client
can compile with:
  ~/EzPC/Athos/CompileONNXGraph.py --config config.json --role client